+++ iti-react/Util/index.ts
@@ -1,2 +1,1 @@
-﻿export * from './FormToObject'
-export * from './RoutingUtil'
+﻿export * from './RoutingUtil'